working on first modeldefinition example
