+++ src/sbmlutils/comp/comp.py
@@ -8,7 +8,7 @@ of the dynamic FBA models.
 """
 
 import logging
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, List, Optional
 
 import libsbml
 
@@ -87,347 +87,15 @@ def get_submodel_frameworks(doc: libsbml.SBMLDocument) -> Dict[str, Any]:
     return frameworks
 
 
-class ExternalModelDefinition(factory.Sbase):
-    """ExternalModelDefinition."""
-
-    def __init__(
-        self,
-        sid: str,
-        source: str,
-        modelRef: str,
-        md5: str = None,
-        name: str = None,
-        sboTerm: str = None,
-        metaId: str = None,
-    ):
-        """Create an ExternalModelDefinition."""
-        super(ExternalModelDefinition, self).__init__(
-            sid=sid, name=name, sboTerm=sboTerm, metaId=metaId
-        )
-        self.source = source
-        self.modelRef = modelRef
-        self.md5 = md5
-
-    def create_sbml(self, model: libsbml.Model) -> libsbml.ExternalModelDefinition:
-        """Create ExternalModelDefinition."""
-        doc = model.getSBMLDocument()
-        cdoc = doc.getPlugin("comp")
-        extdef = cdoc.createExternalModelDefinition()
-        self._set_fields(extdef, model)
-        return extdef
-
-    def _set_fields(
-        self, obj: libsbml.ExternalModelDefinition, model: libsbml.Model
-    ) -> None:
-        """Set fields on ExternalModelDefinition."""
-        super(ExternalModelDefinition, self)._set_fields(obj, model)
-        obj.setModelRef(self.modelRef)
-        obj.setSource(self.source)
-        if self.md5 is not None:
-            obj.setMd5(self.md5)
-
-
-class Submodel(factory.Sbase):
-    """Submodel."""
-
-    def __init__(
-        self,
-        sid: str,
-        modelRef: str = None,
-        timeConversionFactor: str = None,
-        extentConversionFactor: str = None,
-        name: str = None,
-        sboTerm: str = None,
-        metaId: str = None,
-    ):
-        """Create a Submodel."""
-        super(Submodel, self).__init__(
-            sid=sid, name=name, sboTerm=sboTerm, metaId=metaId
-        )
-        self.modelRef = modelRef
-        self.timeConversionFactor = timeConversionFactor
-        self.extentConversionFactor = extentConversionFactor
-
-    def create_sbml(self, model: libsbml.Model) -> libsbml.Submodel:
-        """Create SBML Submodel."""
-        cmodel = model.getPlugin("comp")
-        submodel = cmodel.createSubmodel()
-        self._set_fields(submodel, model)
-
-        submodel.setModelRef(self.modelRef)
-        if self.timeConversionFactor:
-            submodel.setTimeConversionFactor(self.timeConversionFactor)
-        if self.extentConversionFactor:
-            submodel.setExtentConversionFactor(self.extentConversionFactor)
-
-        return submodel
-
-    def _set_fields(self, obj: libsbml.Submodel, model: libsbml.Model) -> None:
-        super(Submodel, self)._set_fields(obj, model)
-
-
-class SbaseRef(factory.Sbase):
-    """SBaseRef."""
-
-    def __init__(
-        self,
-        sid: str,
-        portRef: Optional[str] = None,
-        idRef: Optional[str] = None,
-        unitRef: Optional[str] = None,
-        metaIdRef: Optional[str] = None,
-        name: Optional[str] = None,
-        sboTerm: Optional[str] = None,
-        metaId: Optional[str] = None,
-    ):
-        """Create an SBaseRef."""
-        super(SbaseRef, self).__init__(
-            sid=sid, name=name, sboTerm=sboTerm, metaId=metaId
-        )
-        self.portRef = portRef
-        self.idRef = idRef
-        self.unitRef = unitRef
-        self.metaIdRef = metaIdRef
-
-    def _set_fields(self, obj: libsbml.SBaseRef, model: libsbml.Model) -> None:
-        super(SbaseRef, self)._set_fields(obj, model)
-
-        obj.setId(self.sid)
-        if self.portRef is not None:
-            obj.setPortRef(self.portRef)
-        if self.idRef is not None:
-            obj.setIdRef(self.idRef)
-        if self.unitRef is not None:
-            unit_str = factory.Unit.get_unit_string(self.unitRef)
-            obj.setUnitRef(unit_str)
-        if self.metaIdRef is not None:
-            obj.setMetaIdRef(self.metaIdRef)
-
-
-class ReplacedElement(SbaseRef):
-    """ReplacedElement."""
-
-    def __init__(
-        self,
-        sid: str,
-        elementRef: str,
-        submodelRef: str,
-        deletion: Optional[str] = None,
-        conversionFactor: Optional[str] = None,
-        portRef: Optional[str] = None,
-        idRef: Optional[str] = None,
-        unitRef: Optional[str] = None,
-        metaIdRef: Optional[str] = None,
-        name: Optional[str] = None,
-        sboTerm: Optional[str] = None,
-        metaId: Optional[str] = None,
-    ):
-        """Create a ReplacedElement."""
-        super(ReplacedElement, self).__init__(
-            sid=sid,
-            portRef=portRef,
-            idRef=idRef,
-            unitRef=unitRef,
-            metaIdRef=metaIdRef,
-            name=name,
-            sboTerm=sboTerm,
-            metaId=metaId,
-        )
-        self.elementRef = elementRef
-        self.submodelRef = submodelRef
-        self.deletion = deletion
-        self.conversionFactor = conversionFactor
-
-    def create_sbml(self, model: libsbml.Model) -> libsbml.ReplacedElement:
-        """Create SBML ReplacedElement."""
-        # resolve port element
-        e = model.getElementBySId(self.elementRef)
-        if not e:
-            # fallback to units (only working if no name shadowing)
-            e = model.getUnitDefinition(self.elementRef)
-            if not e:
-                raise ValueError(
-                    f"Neither SBML element nor UnitDefinition found for elementRef: "
-                    f"'{self.elementRef}' in '{self}'"
-                )
-
-        eplugin = e.getPlugin("comp")
-        obj = eplugin.createReplacedElement()
-        self._set_fields(obj, model)
-
-        return obj
-
-    def _set_fields(self, obj: libsbml.ReplacedElement, model: libsbml.Model) -> None:
-        super(ReplacedElement, self)._set_fields(obj, model)
-        obj.setSubmodelRef(self.submodelRef)
-        if self.deletion:
-            obj.setDeletion(self.deletion)
-        if self.conversionFactor:
-            obj.setConversionFactor(self.conversionFactor)
-
-
-class ReplacedBy(SbaseRef):
-    """ReplacedBy."""
-
-    def __init__(
-        self,
-        sid: str,
-        elementRef: str,
-        submodelRef: str,
-        portRef: Optional[str] = None,
-        idRef: Optional[str] = None,
-        unitRef: Optional[str] = None,
-        metaIdRef: Optional[str] = None,
-        name: Optional[str] = None,
-        sboTerm: Optional[str] = None,
-        metaId: Optional[str] = None,
-    ):
-        """Create a ReplacedElement."""
-        super(ReplacedBy, self).__init__(
-            sid=sid,
-            portRef=portRef,
-            idRef=idRef,
-            unitRef=unitRef,
-            metaIdRef=metaIdRef,
-            name=name,
-            sboTerm=sboTerm,
-            metaId=metaId,
-        )
-        self.elementRef = elementRef
-        self.submodelRef = submodelRef
-
-    def create_sbml(
-        self, sbase: libsbml.SBase, model: libsbml.Model
-    ) -> libsbml.ReplacedBy:
-        """Create SBML ReplacedBy."""
-        sbase_comp: libsbml.CompSBasePlugin = sbase.getPlugin("comp")
-        rby: libsbml.ReplacedBy = sbase_comp.createReplacedBy()
-        self._set_fields(rby, model)
-
-        return rby
-
-    def _set_fields(self, rby: libsbml.ReplacedBy, model: libsbml.Model) -> None:
-        """Set fields in ReplacedBy."""
-        super(ReplacedBy, self)._set_fields(rby, model)
-        rby.setSubmodelRef(self.submodelRef)
-
-
-class Deletion(SbaseRef):
-    """Deletion."""
-
-    def __init__(
-        self,
-        sid: str,
-        submodelRef: str,
-        portRef: Optional[str] = None,
-        idRef: Optional[str] = None,
-        unitRef: Optional[str] = None,
-        metaIdRef: Optional[str] = None,
-        name: Optional[str] = None,
-        sboTerm: Optional[str] = None,
-        metaId: Optional[str] = None,
-    ):
-        """Initialize Deletion."""
-        super(Deletion, self).__init__(
-            sid=sid,
-            portRef=portRef,
-            idRef=idRef,
-            unitRef=unitRef,
-            metaIdRef=metaIdRef,
-            name=name,
-            sboTerm=sboTerm,
-            metaId=metaId,
-        )
-        self.submodelRef = submodelRef
-
-    def create_sbml(self, model: libsbml.Model) -> libsbml.Deletion:
-        """Create SBML Deletion."""
-        cmodel: libsbml.CompModelPlugin = model.getPlugin("comp")
-        submodel: libsbml.Submodel = cmodel.getSubmodel(self.submodelRef)
-        deletion: libsbml.Deletion = submodel.createDeletion()
-        self._set_fields(deletion, model)
-
-        return deletion
-
-    def _set_fields(self, obj: libsbml.Deletion, model: libsbml.Model) -> None:
-        """Set fields on Deletion."""
-        super(Deletion, self)._set_fields(obj, model)
-
-
-##########################################################################
-# Ports
-##########################################################################
-# Ports are stored in an optional child ListOfPorts object, which, if
-# present, must contain one or more Port objects.  All of the Ports
-# present in the ListOfPorts collectively define the 'port interface' of
-# the Model.
-PORT_TYPE_PORT = "port"
-PORT_TYPE_INPUT = "input port"
-PORT_TYPE_OUTPUT = "output port"
-
-
-class Port(SbaseRef):
-    """Port."""
-
-    def __init__(
-        self,
-        sid: str,
-        portRef: Optional[str] = None,
-        idRef: Optional[str] = None,
-        unitRef: Optional[str] = None,
-        metaIdRef: Optional[str] = None,
-        portType: Optional[str] = PORT_TYPE_PORT,
-        name: Optional[str] = None,
-        sboTerm: Optional[str] = None,
-        metaId: Optional[str] = None,
-    ):
-        """Create a Port."""
-        super(Port, self).__init__(
-            sid=sid,
-            portRef=portRef,
-            idRef=idRef,
-            unitRef=unitRef,
-            metaIdRef=metaIdRef,
-            name=name,
-            sboTerm=sboTerm,
-            metaId=metaId,
-        )
-        self.portType = portType
-
-    def create_sbml(self, model: libsbml.Model) -> libsbml.Port:
-        """Create SBML for Port."""
-        cmodel = model.getPlugin("comp")
-        p = cmodel.createPort()
-        self._set_fields(p, model)
-
-        if self.sboTerm is None:
-            if self.portType == PORT_TYPE_PORT:
-                # SBO:0000599 - port
-                sbo = 599
-            elif self.portType == PORT_TYPE_INPUT:
-                # SBO:0000600 - input port
-                sbo = 600
-            elif self.portType == PORT_TYPE_OUTPUT:
-                # SBO:0000601 - output port
-                sbo = 601
-            p.setSBOTerm(sbo)
-
-        return p
-
-    def _set_fields(self, obj: libsbml.Port, model: libsbml.Model) -> None:
-        """Set fields on Port."""
-        super(Port, self)._set_fields(obj, model)
-
-
 def create_ports(
     model: libsbml.Model,
     portRefs: Optional[Any] = None,
     idRefs: Optional[Any] = None,
     unitRefs: Optional[Any] = None,
     metaIdRefs: Optional[Any] = None,
-    portType: str = PORT_TYPE_PORT,
+    portType: str = factory.PORT_TYPE_PORT,
     suffix: str = "_port",
-) -> List[Port]:
+) -> List[factory.Port]:
     """Create ports for given model.
 
     Helper function to create port creation.
@@ -479,7 +147,7 @@ def _create_port(
     idRef: Optional[str] = None,
     unitRef: Optional[str] = None,
     metaIdRef: Optional[str] = None,
-    portType: str = PORT_TYPE_PORT,
+    portType: str = factory.PORT_TYPE_PORT,
 ) -> libsbml.Port:
     """Create port in given model."""
     cmodel: libsbml.CompModelPlugin = model.getPlugin("comp")
@@ -503,13 +171,13 @@ def _create_port(
         ref = metaIdRef
     if name is None and ref is not None:
         p.setName(f"port {ref}")
-    if portType == PORT_TYPE_PORT:
+    if portType == factory.PORT_TYPE_PORT:
         # SBO:0000599 - port
         p.setSBOTerm(599)
-    elif portType == PORT_TYPE_INPUT:
+    elif portType == factory.PORT_TYPE_INPUT:
         # SBO:0000600 - input port
         p.setSBOTerm(600)
-    elif portType == PORT_TYPE_OUTPUT:
+    elif portType == factory.PORT_TYPE_OUTPUT:
         # SBO:0000601 - output port
         p.setSBOTerm(601)
